finished and tested reduction (square norm) for float

+++ code/batteryviz/src/BatteryApp.h
@@ -84,7 +84,7 @@ protected:
 
 	blib::DiffusionSolver<double> _diffSolver;
 	blib::MultigridSolver<double> _multiSolver;
-	blib::MultigridGPU<double> _multigridGPUSolver;
+	blib::MultigridGPU<float> _multigridGPUSolver;
 
 
 	friend Ui;
